+++ popperJs/src/modules/popperjsPositioner.module.ts
@@ -1,0 +1,21 @@
+import {NgModule} from "@angular/core";
+
+import {PopperJsPositionerComponent} from '../plugins/positioner/popperJs/popperJsPositioner.component';
+
+/**
+ * Module for popperjs positioner plugin
+ */
+@NgModule(
+{
+    declarations:
+    [
+        PopperJsPositionerComponent
+    ],
+    exports:
+    [
+        PopperJsPositionerComponent
+    ]
+})
+export class PopperJsPositionerModule
+{
+}