+++ popperJs/src/plugins/positioner/components.ts
@@ -1,0 +1,1 @@
+export * from './popperJs/popperJsPositioner.component';