+++ popperJs/src/plugins/positioner/popperJs/popperJsPositioner.interface.ts
@@ -1,0 +1,15 @@
+import {Positioner} from '@anglr/select';
+
+/**
+ * Popper js positioner options
+ */
+export interface PopperJsPositionerOptions
+{
+}
+
+/**
+ * Public API for 'PopperJsPositionerComponent'
+ */
+export interface PopperJsPositioner extends Positioner
+{
+}